Fix: navbar corrected
