+++ src/App.jsx
@@ -1,6 +1,6 @@
 import { Routes, Route } from "react-router-dom";
 
-import { Home, Error, Profile, Login, SignUp } from "./Pages";
+import { Home, Error, Profile } from "./Pages";
 
 import { Navbar, Footer } from "./Components";
 
@@ -11,8 +11,6 @@ const App = () => {
       <Routes>
         <Route path="/" element={<Home />} />
         <Route path="/profile" element={<Profile />} />
-        <Route path="/login" element={<Login />} />
-        <Route path="/signup" element={<SignUp />} />
         {/* <Route path="/posts/: id" element={<PostDetails />} /> */}
         <Route path="*" element={<Error />} />
       </Routes>
