+++ src/Components/Modal/Login.module.scss
@@ -9,9 +9,8 @@
   align-items: center;
 
   .innerBox {
-    min-width: 480px;
+    width: 500px;
     height: fit-content;
-    width: fit-content;
     background-color: #fff;
     box-shadow: 1px 1px 4px rgba(0, 0, 0, 0.2);
     padding: 30px;
@@ -80,6 +79,6 @@
 
 @media only screen and (max-width: 600px) {
   .innerBox {
-    min-width: 300px;
+    width: 380px !important;
   }
 }
