+++ src/Components/Modal/SignUp.jsx
@@ -1,64 +1,38 @@
+/* eslint-disable jsx-a11y/anchor-is-valid */
 /* eslint-disable jsx-a11y/no-noninteractive-element-interactions */
 import React, { useState } from "react";
-import { Link } from "react-router-dom";
 import InputControl from "./InputControl";
 import styles from "./SignUp.module.scss";
-
-// import { auth } from "../../firebase";
-// import {motion} from "framer-motion"
-// import { createUserWithEmailAndPassword, updateProfile} from "firebase/auth";
-
 const SignUp = (props) => {
-  //   const history=useHistory();
   const { setSignup, setOpenLogin } = props;
-  //   const{setCurrentUser}
-  //   const navigate = useNavigate();
   const [show, setShow] = useState("show");
   const [values, setValues] = useState({
     name: "",
     email: "",
     pass: "",
+    loc: "",
+    ComName: "",
+    Desig: "",
   });
-  //   const [errorMsg, setErrorMsg] = useState("");
-  //   const [submitButtonDisabled, setSubmitButtonDisabled] = useState(false);
+  const [errorMsg, setErrorMsg] = useState("");
+  const [submitButtonDisabled, setSubmitButtonDisabled] = useState(false);
 
-  //   const handleSubmission = () => {
-  //     ////validation for ece.nits.ac.in
-  //       let atpos = values.email.indexOf("@");
-  //       let domain = values.email.split("@")[1];
-  //     if(atpos<1 || domain!=="ece.nits.ac.in"){
-  //       setErrorMsg("Only ECE valid");
-  //       return;
-  //     }
-  //     if (!values.name || !values.email || !values.pass) {
-  //       setErrorMsg("Fill all fields");
-  //       return;
-  //     }
-  //     setErrorMsg("");
-
-  //     setSubmitButtonDisabled(true);
-  //     /////creating new user after validating
-  //     createUserWithEmailAndPassword(auth, values.email, values.pass)
-  //       .then(async (res) => {
-  //         setSubmitButtonDisabled(false);
-  //         const user = res.user;
-  //         await updateProfile(user, {
-  //           displayName: values.name,
-  //         });
-  //         setCurrentUser(auth.currentUser);
-  //         setSignup(false);
-  //         navigate("/");
-  //       })
-  //       .catch((err) => {
-  //         setSubmitButtonDisabled(false);
-  //         if (err.code === 'auth/email-already-in-use') {
-  //             setErrorMsg('Email Already in Use');
-  //           }else{
-  //             setErrorMsg(err.message);
-  //           }
-
-  //       });
-  //   };
+  const handleSubmission = () => {
+    if (
+      !values.name ||
+      !values.email ||
+      !values.pass ||
+      !values.ComName ||
+      !values.Desig ||
+      !values.loc
+    ) {
+      setErrorMsg("Fill all fields");
+      return;
+    }
+    console.log(values);
+    setErrorMsg("");
+    setSubmitButtonDisabled(true);
+  };
   const handleClick = () => {
     setSignup(false);
     setOpenLogin(true);
@@ -98,47 +72,43 @@ const SignUp = (props) => {
           <div className={styles.fieldsChild}>
             <InputControl
               label="Location"
-              placeholder="Enter Location"
+              placeholder="Location"
               onChange={(event) =>
                 setValues((prev) => ({ ...prev, loc: event.target.value }))
               }
-              type={show !== "show" ? "text" : "password"}
+              type="text"
             />
             <InputControl
               label="Company Name"
-              placeholder="Enter your company name"
+              placeholder="Company name"
               onChange={(event) =>
                 setValues((prev) => ({ ...prev, ComName: event.target.value }))
               }
-              type={show !== "show" ? "text" : "password"}
+              type="text"
             />
             <InputControl
               label="Designation"
-              placeholder="Enter your Designation"
+              placeholder="Designation"
               onChange={(event) =>
                 setValues((prev) => ({ ...prev, Desig: event.target.value }))
               }
-              type={show !== "show" ? "text" : "password"}
+              type="text"
             />
           </div>
         </div>
         {values.pass.length > 1 && (
           <p
+            className={styles.showPassword}
             onClick={() => setShow(show === "show" ? "Hide" : "show")}
-            style={{
-              position: "absolute",
-              top: "41%",
-              fontSize: "14px",
-              textDecoration: "underline",
-              cursor: "pointer",
-            }}
           >
             {show} password
           </p>
         )}
         <div className={styles.footer}>
-          {/* <b className={styles.error}>{errorMsg}<z/b> */}
-          <button>Signup</button>
+          <b className={styles.error}>{errorMsg}</b>
+          <button onClick={handleSubmission} disabled={submitButtonDisabled}>
+            Signup
+          </button>
           <button
             className={styles.cancel}
             onClick={() => {
@@ -150,9 +120,7 @@ const SignUp = (props) => {
           <p>
             Already have an account?{" "}
             <span>
-              <Link to="/login" onClick={handleClick}>
-                Login
-              </Link>
+              <a onClick={handleClick}>Login</a>
             </span>
           </p>
         </div>
