+++ src/Components/Modal/SignUp.module.scss
@@ -9,7 +9,7 @@
   align-items: center;
 
   .innerBox {
-    min-width: 480px;
+    width: 500px;
     height: fit-content;
     width: fit-content;
     background-color: #fff;
@@ -19,13 +19,14 @@
     display: flex;
     flex-direction: column;
     gap: 30px;
-
+    position: relative;
     .fields {
       display: flex;
       flex-direction: row;
-      gap: 2.5rem;
+      gap: 2rem;
 
       .fieldsChild {
+        width: 100%;
         display: flex;
         flex-direction: column;
         gap: 1rem;
@@ -89,9 +90,23 @@
     }
   }
 }
-
+.showPassword {
+  position: absolute;
+  top: 61%;
+  font-size: 14px;
+  text-decoration: underline;
+  cursor: pointer;
+}
 @media only screen and (max-width: 600px) {
   .innerBox {
     min-width: 300px;
+    gap: 0px !important;
+  }
+  .fields {
+    flex-direction: column !important;
+    gap: 0.5rem !important;
+  }
+  .showPassword {
+    top: 74% !important ;
   }
 }
