+++ src/Pages/index.jsx
@@ -1,7 +1,5 @@
 import Home from "./Home/Home";
 import Error from "./Error/Error";
 import Profile from "./Profile/Profile";
-import Login from "./Modal/Login";
-import SignUp from "./Modal/SignUp";
 
-export { Home, Error, Profile, Login, SignUp };
+export { Home, Error, Profile };
